Add moon crate.
