+++ crates/cli/src/commands/bin.rs
@@ -1,5 +1,6 @@
-use crate::helpers::{load_workspace, AnyError};
+use crate::helpers::AnyError;
 use clap::ValueEnum;
+use moon::load_workspace;
 use moon_terminal::safe_exit;
 use moon_toolchain::RuntimeTool;
