+++ crates/cli/src/commands/check.rs
@@ -1,5 +1,6 @@
 use crate::commands::run::{run_target, RunOptions};
-use crate::helpers::{generate_project_graph, load_workspace, AnyError};
+use crate::helpers::AnyError;
+use moon::{generate_project_graph, load_workspace};
 use moon_logger::trace;
 use moon_project::Project;
 use std::env;
