+++ crates/cli/src/commands/docker/prune.rs
@@ -1,6 +1,7 @@
 use crate::commands::docker::scaffold::DockerManifest;
-use crate::helpers::{generate_project_graph, load_workspace_with_toolchain, AnyError};
+use crate::helpers::AnyError;
 use futures::future::try_join_all;
+use moon::{generate_project_graph, load_workspace_with_toolchain};
 use moon_config::ProjectLanguage;
 use moon_node_lang::{PackageJson, NODE};
 use moon_project_graph::ProjectGraph;
