+++ crates/cli/src/commands/node/run_script.rs
@@ -1,4 +1,5 @@
-use crate::helpers::{generate_project_graph, load_workspace_with_toolchain, AnyError};
+use crate::helpers::AnyError;
+use moon::{generate_project_graph, load_workspace_with_toolchain};
 use moon_error::MoonError;
 use std::env;
