+++ crates/cli/src/commands/teardown.rs
@@ -1,4 +1,5 @@
-use crate::helpers::{create_progress_bar, load_workspace, AnyError};
+use crate::helpers::{create_progress_bar, AnyError};
+use moon::load_workspace;
 
 pub async fn teardown() -> Result<(), AnyError> {
     let done = create_progress_bar("Tearing down toolchain and uninstalling tools...");
