+++ crates/cli/src/helpers.rs
@@ -1,81 +1,11 @@
 use console::{set_colors_enabled, set_colors_enabled_stderr};
 use indicatif::{ProgressBar, ProgressStyle};
-use moon_config::PlatformType;
-use moon_dep_graph::DepGraphBuilder;
 use moon_logger::color::{no_color, supports_color};
-use moon_node_platform::NodePlatform;
-use moon_project::ProjectError;
-use moon_project_graph::{ProjectGraph, ProjectGraphBuilder};
-use moon_system_platform::SystemPlatform;
 use moon_terminal::create_theme;
-use moon_workspace::{Workspace, WorkspaceError};
-use rustc_hash::FxHashMap;
 use std::env;
 use std::time::Duration;
-use strum::IntoEnumIterator;
 
 pub type AnyError = Box<dyn std::error::Error>;
-
-/// Loads the workspace and registers all available platforms!
-pub async fn load_workspace() -> Result<Workspace, WorkspaceError> {
-    let mut workspace = Workspace::load().await?;
-
-    workspace.register_platform(Box::new(SystemPlatform::default()));
-
-    if let Some(node_config) = &workspace.toolchain.config.node {
-        workspace.register_platform(Box::new(NodePlatform::new(node_config, &workspace.root)));
-    }
-
-    workspace.signin_to_moonbase().await?;
-
-    Ok(workspace)
-}
-
-// Some commands require the toolchain to exist, but don't use
-// the action runner. This is a simple flow to wire up the tools.
-pub async fn load_workspace_with_toolchain() -> Result<Workspace, WorkspaceError> {
-    let mut workspace = load_workspace().await?;
-    let mut last_versions = FxHashMap::default();
-
-    // Use exhaustive checks so we don't miss a platform
-    for platform in PlatformType::iter() {
-        match platform {
-            PlatformType::Node => {
-                if let Some(node_config) = &workspace.toolchain.config.node {
-                    workspace
-                        .toolchain
-                        .node
-                        .setup(&node_config.version, &mut last_versions)
-                        .await?;
-                }
-            }
-            PlatformType::System | PlatformType::Unknown => {}
-        }
-    }
-
-    Ok(workspace)
-}
-
-pub fn build_dep_graph<'g>(
-    workspace: &'g Workspace,
-    project_graph: &'g ProjectGraph,
-) -> DepGraphBuilder<'g> {
-    DepGraphBuilder::new(&workspace.platforms, project_graph)
-}
-
-pub async fn generate_project_graph(
-    workspace: &mut Workspace,
-) -> Result<ProjectGraph, ProjectError> {
-    let mut builder = ProjectGraphBuilder {
-        cache: &workspace.cache,
-        config: &workspace.projects_config,
-        platforms: &mut workspace.platforms,
-        workspace_config: &workspace.config,
-        workspace_root: &workspace.root,
-    };
-
-    Ok(builder.build().await?)
-}
 
 pub fn create_progress_bar<S: AsRef<str>, F: AsRef<str>>(start: S) -> impl FnOnce(F, bool) {
     let pb = ProgressBar::new_spinner();
